backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0  # For SQLite support (no installation required)

# PostgreSQL support (optional - only install if using PostgreSQL)
# psycopg2-binary==2.9.9
# asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# NLP and ML (optional - only needed for production NLP model)
# transformers==4.35.2
# torch>=2.0.0
# sentencepiece==0.1.99
# scikit-learn==1.3.2
# nltk==3.8.1

# Redis for caching (optional)
# redis==5.0.1
# hiredis==2.2.3

# HTTP Client
httpx==0.25.2
requests==2.31.0

# LLM Integration
openai>=1.14.0

# Data Processing (optional - only for analytics)
# pandas==2.1.4
# numpy==1.26.2
datasets>=2.20.0
# datasets>=2.19.0  # For BANKING77 retraining augmentation

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Monitoring and Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Email (optional)
# fastapi-mail==1.4.1

# Date and Time
python-dateutil==2.8.2
pytz==2023.3
